+++ CHANGELOG.md
@@ -19,6 +19,9 @@ CWD: current working directory
 
 ### Added 
 
+- Warning is shown for having a file with invalid UTF8 content within a template 
+  after initializing a project for the template in question. 
+  A file with invalid utf8 content is just copied and not augmented.
 - In the console prompt: Default value is also shown if no value to a key is given.
 - Added 2 ClI/environment option: ( left_delimiter, right_delimiter )
   Allows to change what determines the left and right border of a placeholder in file while initializing.